third_party/WebKit/LayoutTests/storage/indexeddb/version-change-abort-expected.txt

CONSOLE ERROR: line 58: Uncaught Error: This should not be caught
CONSOLE ERROR: line 1: AbortError: Version change transaction was aborted in upgradeneeded event handler.
Ensure that aborted VERSION_CHANGE transactions are completely rolled back

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


dbname = "version-change-abort.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
PASS trans instanceof IDBTransaction is true
db.createObjectStore('store1')
openRequest1 complete
PASS db.version is 1

vcreq = indexedDB.open(dbname, 2)
openRequest2() callback
PASS db.version is 2
PASS vcreq.transaction instanceof IDBTransaction is true
db.deleteObjectStore('store1')
db.createObjectStore('store2')
raising exception
PASS Got expected error: 'Uncaught Error: This should not be caught'
PASS sawGlobalErrorUncaughtException is true

openRequest2Error() callback
PASS sawGlobalErrorUncaughtException is true
PASS sawTransactionAbort is true
PASS Got expected error: 'AbortError: Version change transaction was aborted in upgradeneeded event handler.'

Verify rollback:
request = indexedDB.open(dbname)
PASS db.version is 1
PASS db.objectStoreNames.contains('store1') is true
PASS db.objectStoreNames.contains('store2') is false
PASS successfullyParsed is true

TEST COMPLETE
